+++ projects/experimental/directives/card/card.module.ts
@@ -1,6 +1,7 @@
 import {NgModule} from '@angular/core';
 
-import {TuiCardComponent, TuiCardDirective} from './card.directive';
+import {TuiCardComponent} from './card.component';
+import {TuiCardDirective} from './card.directive';
 
 @NgModule({
     declarations: [TuiCardComponent, TuiCardDirective],
