+++ projects/experimental/directives/card/index.ts
@@ -1,2 +1,3 @@
+export * from './card.component';
 export * from './card.directive';
 export * from './card.module';
